Extract CompetitionsViewController test helpers to a separate file
